# feat: implement state-of-the-art CRISP-based prompt engineering (Phase 1 complete)