--- a/.gitignore
+++ b/.gitignore
@@ -44,3 +44,7 @@ TODO.md
 *-plan.md
 *-task.md
 dist/
+thinktank_*/
+REVIEW-CONTEXT.md
+PLAN-CONTEXT.md
+TICKET-CONTEXT.md