--- a/src/api.ts
+++ b/src/api.ts
@@ -5,6 +5,8 @@
  * Following radical simplification philosophy.
  */
 
+import { CORE_EXAMPLES } from "./prompting/examples.js";
+
 /**
  * Create a structured log entry with timestamp, level, and message.
  *
@@ -32,11 +34,16 @@ function logStructured(
 }
 
 /**
- * Enhanced CRISP-based elevation prompt for transforming user requests.
+ * Build the enhanced CRISP-based elevation prompt dynamically.
  * Uses proven prompt engineering techniques including structured formatting,
- * few-shot examples, and explicit output constraints.
+ * few-shot examples from the curated library, and explicit output constraints.
  */
-const ELEVATION_PROMPT = `<role>Expert prompt engineer specializing in technical communication</role>
+function buildElevationPrompt(): string {
+  const examplesSection = CORE_EXAMPLES.map(
+    (example) => `\nInput: "${example.input}"\nOutput: "${example.output}"`,
+  ).join("");
+
+  return `<role>Expert prompt engineer specializing in technical communication</role>
 
 <context>Transform requests using proven CRISP structure (Context, Role, Instructions, Specifics, Parameters)</context>
 
@@ -48,28 +55,21 @@ const ELEVATION_PROMPT = `<role>Expert prompt engineer specializing in technical
 5. Specify success criteria and validation methods
 </instructions>
 
-<examples>
-Input: "help with my code"
-Output: "Review this [LANGUAGE] codebase for performance bottlenecks. Focus on: 1) Inefficient algorithms (O(n²)+), 2) Memory leaks, 3) Unnecessary operations. Provide specific line numbers and optimization suggestions. Format: Markdown with code snippets."
-
-Input: "write about AI"
-Output: "Create a 1200-word technical analysis of AI applications in [DOMAIN]. Structure: Executive summary (200w), Current state (400w), Emerging trends (400w), Implementation considerations (200w). Target audience: Technical decision-makers. Include 3+ case studies."
-
-Input: "analyze this data"
-Output: "Perform statistical analysis on [DATA_TYPE] dataset. Methodology: 1) Descriptive statistics with outlier detection, 2) Trend analysis using [METHOD], 3) Correlation matrix for key variables. Deliverables: Executive summary, detailed findings with visualizations, recommendations with confidence intervals. Format: PDF report with appendices."
-
-Input: "fix my bug"
-Output: "Debug [ERROR_TYPE] in [COMPONENT] (line X). Reproduce steps: [STEPS]. Expected vs actual behavior: [COMPARISON]. Provide: 1) Root cause analysis, 2) Fix with explanation, 3) Prevention strategy, 4) Test cases."
-
-Input: "create a design"
-Output: "Design [ARTIFACT] for [AUDIENCE]. Requirements: [SPECIFICATIONS]. Constraints: [TECHNICAL/BUSINESS_LIMITS]. Deliverables: 1) Wireframes/mockups, 2) Technical specifications, 3) Implementation timeline, 4) Success metrics. Format: Design system with documentation."
+<examples>${examplesSection}
 </examples>
 
 <output_constraints>
 Output ONLY the transformed prompt. No explanations, headers, or meta-commentary.
 </output_constraints>
 
-Transform this request:`;
+Transform this request:`.trim();
+}
+
+/**
+ * Enhanced CRISP-based elevation prompt for transforming user requests.
+ * Dynamically constructed from curated examples to ensure consistency.
+ */
+const ELEVATION_PROMPT = buildElevationPrompt();
 
 /**
  * Elevate a user prompt using direct Gemini API call.