… SYNTHESIS.md
… TODO.md
… docs/PERFORMANCE_BASELINE.md
… src/__snapshots__/api.golden.test.ts.snap
… src/api.golden.test.ts
… src/prompting/examples.test.ts
… test/cli.integration.test.ts